Iterate over PRs in desc order
Previously we found a merged PR matching the latest PyPi release
and then found a newer merged PR and tried to do a github/pypi/fedora release
of that version. If that PR was for some reason wrong, the bot
would crash loop on that PR forever.

With this change we simply find the newest merged release PR
and try to do a github/pypi/fedora release of that version.
Even if the PR is bad, the bot returns to the Light side of the Force
once the maintainer creates newer (correct) PR.

Fixes #45